requirements.txt
amqp==2.5.2
args==0.1.0
attrs==19.3.0
awesome-slugify==1.6.5
beautifulsoup4==4.6.0
billiard==3.5.0.5
biopython==1.76
bs4==0.0.1
celery==4.1.0
certifi==2019.11.28
chardet==3.0.4
clint==0.5.1
coloredlogs==10.0
cssselect==1.1.0
dataclasses==0.6
docopt==0.6.2
eutils==0.6.0
filelock==3.0.12
habanero==0.7.2
humanfriendly==4.18
idna==2.6
importlib-metadata==1.3.0
inflection==0.3.1
jsonpointer==1.12
kombu==4.6.7
lxml==4.2.5
metapub==0.5.3
more-itertools==8.0.2
numpy==1.17.4
packaging==19.2
pandas==0.25.3
patent-client==1.1.4
pathvalidate==0.16.2
pluggy==0.13.1
pubmed-lookup==0.2.3
py==1.8.0
pymed==0.8.9
pyparsing==2.4.5
pypatent==1.2.0
PyPDF2==1.26.0
pytest==5.3.2
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
redis==2.10.6
regex==2019.12.20
requests==2.22.0
requests-cache==0.5.2
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
tabulate==0.8.6
toml==0.10.0
tox==3.14.2
tqdm==4.41.0
Unidecode==0.4.21
urllib3==1.22
uspto-opendata-python==0.8.3
vine==1.3.0
virtualenv==16.7.9
wcwidth==0.1.7
xmltodict==0.12.0
zipp==0.6.0
